Return remote cmd exit status, update `deploy`
This commit sets up `mcv.remote._exec_command`, which is just like
paramiko.SSHClient.exec_command, but it also returns the command's
exit status.

It then redoes the `mcv.remote` functionality in terms of this helper,
so it can pass the exit status out to callers.

It then updates the callers to expect getting the exit status.

In the course of all of these updates, this commit also redoes
`mcv.remote.deploy` in terms of the new functionality that's been
added, e.g. the better `execute()` and the fixed verbosity
conventions.