Enable pip install --upgrade
This commit adds a flag to the local and remote `pip` modules that
allow you to run `pip install --upgrade` either locally or remotely.

In the course of the doing this, it consolidates the command
generation a bit, since it's the same command regardless of whether
it's executed locally via `subprocess` or remotely via `mcv.remote`.
Thus we now have one function `mcv.pip._install_cmd` that is used
by both local and remote modules.

Having this functionality allows you to force upgrading `pip` packages
even if the system thinks you already have an existing (old) version.