Document passing private key to mcv.remote.connection
This commit adds a comment explaining how to pass a private key file
to mcv.remote.connection (thus/the same way as passing it to the
underlying Paramiko).

This is helpful when you want to specify a private key.  This is
particularly useful/necessary for connecting to a Vagrant instance.