Avoid path joining in mcv.git.export
Prior to this commit, mcv.git.export would attempt to be "smart" about
constructing paths, by having default assumptions about where exported
snapshots of git repos would go.  This is not a good strategy--let the
caller handle all the logic of where the output is going to go, rather
than having an inflexible assumption, and just put the output where
it's asked to.

This commit removes the path joining logic and just puts the output
where requested.