Enable streaming output from mcv.remote.execute
This command makes it possible to stream stdout/stderr from remote
commands while they're happening.  This is essential from long-running
remote commands, and commands where the you can't tell if the remote
command has hung/failed; you want to see the output as it's happening.

Unfortunately this messes up the old verbosity behavior; you don't
want to have to wait until a command has fail-exited to see its
output.  So now if you want to control verbosity you have to pass in
alternate stdout/stderr handles _beforehand_, to handle the streams of
output that may or may not occur.  This is a fine/normal tradeoff
for now; it works just the same as normal local commands, which need
to make the same tradeoff anyway.