Enable no-opt if mount already mounted
Prior to this commit, a `mount()` call would remount if already
mounted.  This commit makes it so that it literally does nothing if
already mounted, unless `remount=True` flag is passed, in which a
mount will result in a remount.  The operation was already idempotent,
but this makes it sort of more safely side-effect free if already
in the mounted state.