Turn on absolute import everywhere
Now that we have an mcv.os module, we need absolute import turned on
everywhere we're trying to import top-level `os` module, instead of
`mcv.os`.  This was supposed to be resolved in Python 2.7, but _is
not_--manually enabling absolute imports is still required in 2.7,
though it's finally fixed in 3.

(And for better or worse, the default version shipping on Ubuntu
etc. is still 2 not 3).